refactor public task related structs into models module
